align the build and clean environments
VE-29

+++ tests/clean.sh
@@ -2,15 +2,20 @@
 
 set -x
 
-if [ -n "$VAMP_CHANGE_TARGET" ]; then
-  export VAMP_GIT_BRANCH=$VAMP_CHANGE_TARGET
+if [ -z "$VAMP_GIT_ROOT" ]; then
+  export VAMP_GIT_ROOT=$(git remote -v | grep fetch | awk '{ print $2 }' | awk -F '/' '{ print "git@" $3 ":" $4 }')
 fi
 
 if [ -n "$CHANGE_TARGET" ]; then
-  export VAMP_GIT_BRANCH=$CHANGE_TARGET
+  export VAMP_CHANGE_TARGET=$CHANGE_TARGET
+  export VAMP_CHANGE_URL=$CHANGE_URL
   export VAMP_TAG_PREFIX="pr-$(echo $CHANGE_URL | sed -e 's,.*/vamp-docker-images/pull/,,g')-"
-else
+elif [ -n "$BUILD_NUMBER" ]
   export VAMP_TAG_PREFIX="build-$BUILD_NUMBER-"
+fi
+
+if [ -n "$VAMP_CHANGE_TARGET" ]; then
+  export VAMP_GIT_BRANCH=$VAMP_CHANGE_TARGET
 fi
 
 if [ -z "$VAMP_GIT_BRANCH" ]; then
@@ -23,7 +28,7 @@ fi
 
 tag=$(echo $VAMP_GIT_BRANCH | sed 's,/,_,g')
 if [ "$VAMP_GIT_BRANCH" = "master" ]; then
-  tag="katana--"
+  tag=katana
 fi
 tag="${VAMP_TAG_PREFIX}${tag}"
 
@@ -31,9 +36,11 @@ exited_containers=$(docker ps -a -f status=exited -q)
 dead_containers=$(docker ps -a -f status=dead -q)
 test -n "${exited_containers}" -o -n "${dead_containers}" && docker rm ${exited_containers} ${dead_containers}
 
-remote_images=$(docker image ls -f reference="magneticio/vamp*:${tag}*" --format '{{.Repository}}:{{.Tag}}')
-local_images=$(docker image ls -f reference="vamp*:${tag}*" --format '{{.Repository}}:{{.Tag}}')
-test -n "${remote_images}" -o -n "${local_images}" && docker rmi -f ${remote_images} ${local_images}
+if [ "$VAMP_GIT_BRANCH" != "master" ]; then
+  remote_images=$(docker image ls -f reference="magneticio/vamp*:${tag}*" --format '{{.Repository}}:{{.Tag}}')
+  local_images=$(docker image ls -f reference="vamp*:${tag}*" --format '{{.Repository}}:{{.Tag}}')
+  test -n "${remote_images}" -o -n "${local_images}" && docker rmi -f ${remote_images} ${local_images}
+fi
 
 dangling_images=$(docker image ls -f dangling=true -q)
 while [ -n "${dangling_images}" ]; do
